pipeline: move local filesystem I/O into its own file
